Fix player's name not appearing on team join
This happened when someone joined the server and it would say:
* has changed to team 'blue'

And you would do +showscores or check the console to see who has just
connected.

Note: the game does not always show that message when a player joins,
this commit doesn't fix that. Player instances are reused and it probably
has something to do with that, but i don't know.