Make CTF/DOM gamemode autoswitch depend on cvar
New cvar:
* sv_ag_gamemode_auto <0|1> - Default: 1

When set to 1 it will change automatically to CTF or DOM if the map fulfills
the criteria for CTF or DOM. This is the classic behaviour.

You can set it to 0 if you don't have the gamemode switching automatically,
which is useful for servers running HLKZ or other custom games.