Add Visual Studio 2022 project files
Every commit we make, we will check if it builds on both VS2019 and VS2022 as the VS2022 project files are using the new C++ 20 compilers (the Windows SDK version was not changed). If the build fails on VS2019, we will have to re-write said commit for VS2019 and stick with VS2019 until the finished product of VIsual Studio 2022 releases. VS2022 is still in beta, and I doubt any of the Phoenix Software developers will want to use potentially unstable software to work on Aura. The reason I've added the VS2022 files, is to just remain futureproof.