Soften spawn blocking criteria
Players no longer block spawn spots by being right below or above it,
when there's a floor separating the player and the spot. Same when
there's a wall in between. If there's just a small obstacle between
the player and the spot then it depends on whether the player is closer
than 128 units and they can see the head or arms of the spawning player
in that spot

This could lead to having a player spawning in the next 6th spot instead
of the 5 ones that people are used to. That may still happen but it will
be more obvious that a player has blocked a spawn spot